Move modules and submodules of basic_test_class to specific folders
Ref: https://github.com/hl7au/au-ps-inferno/issues/43
