+++ lib/au_ps_inferno/utils/attester/basic_test_attester_identifier_slices.rb
