+++ lib/au_ps_inferno/utils/attester/basic_test_attester_resource.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative 'bundle_decorator'
+require_relative '../bundle_decorator'
 
 module AUPSTestKit
   # Resolves Composition attester.party reference from the scratch bundle and loads attester metadata.
