+++ lib/au_ps_inferno/utils/author/basic_test_author_module.rb
