+++ lib/au_ps_inferno/utils/author/basic_test_author_resource.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative 'bundle_decorator'
+require_relative '../bundle_decorator'
 
 module AUPSTestKit
   # Resolves Composition.author reference from the scratch bundle.
