+++ lib/au_ps_inferno/utils/author/basic_test_author_tests.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative 'basic_test_contants_module'
+require_relative '../basic_test_contants_module'
 
 module AUPSTestKit
   # Inferno tests for Composition.author Must Support.
