+++ lib/au_ps_inferno/utils/subject/basic_test_subject_patient_ms_elements.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative 'basic_test_contants_module'
+require_relative '../basic_test_contants_module'
 
 module AUPSTestKit
   # Patient (subject) Must Support primitives and optional extension slices.
