+++ lib/au_ps_inferno/utils/subject/basic_test_subject_resource.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative 'bundle_decorator'
+require_relative '../bundle_decorator'
 
 module AUPSTestKit
   # Resolves Composition.subject (Patient) from the scratch bundle.
